feat: add restore option (#99)
* Add restore option

Some tests may want to use the natural system environment but then restore it completely after the test (so that any environment variables added during the test get removed).  Adding a 'restore' option allows for this.

* Allow for passing options as first argument

* Fix defect

* Fix default value for restore option

* feat: "restore" option and options as first argument

this feature adds a restore option that if passed will restore the process.env environment when the restore() function is called. This varies from the current "clear" functionality as it does not clear the current environment when mockedEnv is called.

* fix: add error tests and documentation about new features

* feat: add TypeScript defininition

* feat: support for ES6 style default imports